requirements.txt
python-dotenv
google-genai
pydantic
jinja2
owlready2